+++ src/main/java/com/telepathicgrunt/the_bumblezone/modCompat/ModChecker.java
@@ -65,7 +65,7 @@ public class ModChecker
 	private static void printErrorToLogs(String currentModID) {
 		Bumblezone.LOGGER.log(Level.INFO, "------------------------------------------------NOTICE-------------------------------------------------------------------------");
 		Bumblezone.LOGGER.log(Level.INFO, " ");
-		Bumblezone.LOGGER.log(Level.INFO, "ERROR: Something broke when trying to add mod compatibility with" + currentModID + ". Please let The Bumblezone developer (TelepathicGrunt) know about this!");
+		Bumblezone.LOGGER.log(Level.INFO, "ERROR: Something broke when trying to add mod compatibility with " + currentModID + ". Please let The Bumblezone developer (TelepathicGrunt) know about this!");
 		Bumblezone.LOGGER.log(Level.INFO, " ");
 		Bumblezone.LOGGER.log(Level.INFO, "------------------------------------------------NOTICE-------------------------------------------------------------------------");
 	}